Corrupting Fine Tuned Legal BERT
This is a custom algorithm involving a corruption of few train samples i.e., corrupting a certain portion of labels of Train Data. This falls into a semi supervised approach.
For a certain portion of Train Data, we used previous Legal BERT model to calculate the test accuracy which is approximately 0.761, so the residue 0.239 is incorrectly predicted, assuming that corrupting the portion of train samples with the portion which is incorrectly predicted on test data contributes something/ focus more on residue during model re-training yielding substantial outputs, which I name conventionally it as Model Regretting Approach.
We corrupted the 0.239 portion of train samples of our Final Train Data and trained our model. The test accuracy is 0.76775