GAN BERT
When there is a limited quantity of labelled data and a significant number of unlabelled data, GAN with BERT has been effectively used to improve text categorization.
For text categorization, this architecture combines the capabilities of BERT and SS-GAN. The generator creates "fake" instances using a 100-dimensional noise vector generated from a Gaussian Distribution as input. The discriminator is an MLP built on top of BERT that takes as input either a false vector created by the generator or a vector formed by BERT from actual data. The discriminator's final layer is a SoftMax layer that produces a k+1 dimension vector of logits, where k is the number of classes in the dataset. The real data is split into two categories: labelled (L) and unlabelled (U).
The discriminator's goal is to determine whether or not the input is a real instance. It must predict which class the input belongs to if it forecasts the input as an actual instance.
The discriminator loss and generator loss are two competing losses that the training process strives to optimize. The discriminator loss is the product of two losses: supervised and unsupervised. Unsupervised loss measures the error in mistakenly detecting a genuine (unlabelled) example as fake and not recognizing a fake example among the initial k categories, whereas supervised loss measures the error in assigning the wrong class to a real example among the original k categories. The generator loss is a combination of two distinct losses: feature matching and unsupervised loss. Unsupervised loss quantifies the error caused by false instances successfully detected by the discriminator, whereas feature matching loss ensures that the generator produces examples whose intermediate representations presented as input to the discriminator are highly close to the genuine ones.
During training, each class's samples are repeated by a factor of log(2|U|/|L|), ensuring the existence of some labelled instances in each batch and avoiding divergences owing to the adversarial training's unsupervised component. The generator is removed from the design during the inference phase, while the remainder is kept.
In comparison to the typical BERT model, this means there is no additional cost at inference time.
We used 5 epochs, and 10 percent label rate, the test accuracy is 75.8