Fine Tuned Legal BERT on Paraphrased
This approach involves paraphrasing a few test samples. We have taken a few test samples from Test Data and paraphrased them. The necessity of this approach is lack of enough train samples to compensate test accuracy.
We pulled a Legal BERT model (BertForSequenceClassification)  from Hugging Face and trained on Final Train Data which is concatenated with Paraphrased Data samples achieving highest Test Accuracy which is 0.834933